requirements.txt
setuptools==51.1.1
wheel==0.36.2

apache-beam[gcp]==2.26.0
google-cloud-secret-manager==2.1.0
google-cloud-storage==1.35.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pyyaml==5.3.1
requests==2.25.1

pandas==1.2.1
numpy==1.19.5

openpyxl==3.0.6
beam-nuggets==0.17.0